+++ castor/lib/core/src/main/scala/io/iohk/atala/castor/core/repository/DIDOperationRepository.scala
@@ -1,7 +1,6 @@
 package io.iohk.atala.castor.core.repository
 
 import io.iohk.atala.castor.core.model.did.PrismDID
-import zio.*
 
 trait DIDOperationRepository[F[_]] {
   def getConfirmedPublishedDIDOperations(did: PrismDID): F[Unit]