+++ prism-agent/service/server/src/main/scala/io/iohk/atala/api/http/EndpointOutputs.scala
@@ -2,6 +2,7 @@ package io.iohk.atala.api.http
 import sttp.model.StatusCode
 import sttp.tapir.json.zio.jsonBody
 import sttp.tapir.{oneOfVariantValueMatcher, *}
+import sttp.tapir.EndpointOutput.OneOfVariant
 
 object EndpointOutputs {
   private def statusCodeMatcher(
@@ -10,30 +11,42 @@ object EndpointOutputs {
     case ErrorResponse(status, _, _, _, _) if status == statusCode.code => true
   }
 
-  val basicFailures: EndpointOutput[ErrorResponse] =
+  def basicFailuresWith(extraFailures: OneOfVariant[ErrorResponse]*) = {
     oneOf(
-      oneOfVariantValueMatcher(
-        StatusCode.BadRequest,
-        jsonBody[ErrorResponse].description("Invalid request parameters")
-      )(statusCodeMatcher(StatusCode.BadRequest)),
-      oneOfVariantValueMatcher(
-        StatusCode.InternalServerError,
-        jsonBody[ErrorResponse].description("Internal server error")
-      )(statusCodeMatcher(StatusCode.InternalServerError))
+      FailureVariant.badRequest,
+      (FailureVariant.internalServerError +: extraFailures): _*
     )
+  }
 
-  val basicFailuresAndNotFound = oneOf(
-    oneOfVariantValueMatcher(
-      StatusCode.NotFound,
-      jsonBody[ErrorResponse].description("Resource could not be found")
-    )(statusCodeMatcher(StatusCode.NotFound)),
-    oneOfVariantValueMatcher(
+  val basicFailures: EndpointOutput[ErrorResponse] = basicFailuresWith()
+
+  val basicFailuresAndNotFound = basicFailuresWith(FailureVariant.notFound)
+
+  object FailureVariant {
+    val badRequest = oneOfVariantValueMatcher(
       StatusCode.BadRequest,
       jsonBody[ErrorResponse].description("Invalid request parameters")
-    )(statusCodeMatcher(StatusCode.BadRequest)),
-    oneOfVariantValueMatcher(
+    )(statusCodeMatcher(StatusCode.BadRequest))
+
+    val internalServerError = oneOfVariantValueMatcher(
       StatusCode.InternalServerError,
       jsonBody[ErrorResponse].description("Internal server error")
     )(statusCodeMatcher(StatusCode.InternalServerError))
-  )
+
+    val notFound = oneOfVariantValueMatcher(
+      StatusCode.NotFound,
+      jsonBody[ErrorResponse].description("Resource could not be found")
+    )(statusCodeMatcher(StatusCode.NotFound))
+
+    val unprocessableEntity = oneOfVariantValueMatcher(
+      StatusCode.UnprocessableEntity,
+      jsonBody[ErrorResponse].description("Unable to process the request")
+    )(statusCodeMatcher(StatusCode.UnprocessableEntity))
+
+    val conflict = oneOfVariantValueMatcher(
+      StatusCode.Conflict,
+      jsonBody[ErrorResponse].description("Cannot process due to conflict with current state of the resource")
+    )(statusCodeMatcher(StatusCode.Conflict))
+  }
+
 }